Improve admin mode, layout, pages
