+++ shop/src/app/makro/page.tsx
@@ -12,7 +12,7 @@ import { useLocalStorage } from "@/hooks/use-local-storage"
 import { MainNavBar } from "@/components/main-nav-bar"
 
 export default function Makro() {
-  const [balance, setBalance] = useLocalStorage("shop-balance", 5000)
+  const [balance, setBalance] = useLocalStorage("shop-balance", 10000)
   const [cartItems, setCartItems] = useLocalStorage<Record<string, number>>("shop-cart", {})
   const [searchQuery, setSearchQuery] = useState("")
 
@@ -183,6 +183,9 @@ export default function Makro() {
       <MainNavBar
         balance={balance}
         cartItemsCount={Object.values(cartItems).reduce((a: number, b: number) => a + (b as number), 0)}
+        onAdminClick={() => {
+          console.log("Admin button clicked");
+        }}
       />
 
       {/* Makro header */}
