+++ shop/src/components/sonner-provider.tsx
@@ -8,9 +8,11 @@ export function SonnerProvider() {
       position="top-right"
       toastOptions={{
         style: {
-          background: "hsl(var(--background))",
-          color: "hsl(var(--foreground))",
-          border: "1px solid hsl(var(--border))",
+          background: "rgba(255, 255, 255, 0.8)", // Solid black background with slight transparency
+          color: "#000000", 
+          border: "1px solid rgba(255, 255, 255, 0.2)", // Subtle white border
+          borderRadius: "8px", // Optional: Rounded corners
+          padding: "12px", // Optional: Add padding for better spacing
         },
       }}
     />
